fix: fixes performance issue in java when calling parent unions
