+++ src/generators/java/renderers/ClassRenderer.ts
@@ -3,36 +3,19 @@ import {
   ConstrainedDictionaryModel,
   ConstrainedObjectModel,
   ConstrainedObjectPropertyModel,
-  ConstrainedUnionModel,
-  InputMetaModel,
-  Preset
+  ConstrainedUnionModel
 } from '../../../models';
 import { FormatHelpers } from '../../../helpers';
-import { JavaGenerator, JavaOptions } from '../JavaGenerator';
+import { JavaOptions } from '../JavaGenerator';
 import { ClassPresetType } from '../JavaPreset';
 import { unionIncludesBuiltInTypes } from '../JavaConstrainer';
-import { JavaDependencyManager } from '../JavaDependencyManager';
 
 /**
  * Renderer for Java's `class` type
  *
  * @extends JavaRenderer
  */
-export class ClassRenderer<
-  RendererModelType extends ConstrainedObjectModel = ConstrainedObjectModel
-> extends JavaRenderer<RendererModelType> {
-  constructor(
-    options: JavaOptions,
-    generator: JavaGenerator,
-    presets: Array<[Preset, unknown]>,
-    model: RendererModelType,
-    inputModel: InputMetaModel,
-    dependencyManager: JavaDependencyManager,
-    private unions: ConstrainedUnionModel[] | undefined
-  ) {
-    super(options, generator, presets, model, inputModel, dependencyManager);
-  }
-
+export class ClassRenderer extends JavaRenderer<ConstrainedObjectModel> {
   async defaultSelf(): Promise<string> {
     const content = [
       await this.renderProperties(),
@@ -117,17 +100,12 @@ ${this.indent(this.renderBlock(content, 2))}
   private getParentUnions(): ConstrainedUnionModel[] | undefined {
     const parentUnions: ConstrainedUnionModel[] = [];
 
-    if (!this.unions) {
+    if (!this.model.options.parentUnions) {
       return undefined;
     }
 
-    for (const unionModel of this.unions) {
-      if (
-        !unionIncludesBuiltInTypes(unionModel) &&
-        unionModel.union.some(
-          (m) => m.name === this.model.name && m.type === this.model.type
-        )
-      ) {
+    for (const unionModel of this.model.options.parentUnions) {
+      if (!unionIncludesBuiltInTypes(unionModel)) {
         parentUnions.push(unionModel);
       }
     }
